iOS feature parity sprint: 98 files, 60 agents across 3 rounds
## Round 1: UI Screens (20 agents)
- App mode switching (Public/Client/Business) with ModeAwareRootView
- Client mode: Home, Bookings, Messages screens + ClientTabView
- Business mode: Home, Schedule, Clients, Inbox screens + BusinessTabView
- Profile: ProfileCardView, AppModeSwitcher, StatsCards
- Social: SocialHubScreenV2 with 5 sub-screens
- Discovery: FilterChips, SortBar, ServiceProviderCard, ServicesCategoryGrid
- Components: ServiceCardView, BusinessWelcomeDialog

## Round 2: Core Fixes (20 agents)
- Walk Firestore save to users/{uid}/walks/ (was broken)
- Badge awarding: removed #if false, fixed type mismatches
- Points formula: aligned tiered bonuses + anti-gaming with Android
- GPS filter: added jitter gate, adaptive accuracy, haptic milestones
- League assignment: ensureLeagueAssignment + processWeekEnd
- Feed: filtering (ForYou/Nearby/Following), pagination, photo posts
- Chat: photo messages, message reactions, delete conversation
- Friend requests: send/accept/reject flow
- Planning: distance calc, loop close, saved walks CRUD
- Share cards: 1080x1350 bitmap, all destinations wired
- Milestones: aligned all categories with Android
- Charity: distance-to-points, monthly aggregation, leaderboard
- Booking model + repository + service types
- Business Firestore: ViewModel with online toggle, earnings, jobs
- Away mode: Firestore persistence, holiday mode, pin/archive

## Round 3: 95% Parity Push (20 agents)
- CoreMotion auto-pause (5s) + auto-resume + GPS drift rejection
- Challenge progress wired to walk completion + default challenges
- Dog reels in feed + functional stories with 24hr expiry
- Share card dog photo hero + PB bitmap + report card sharing
- Directions API routing (OSRM) + execute saved walks
- Scan key: QR camera + key management + Firestore
- Walk console: execute jobs with tracking + earnings
- Client detail: 5-tab screen with notes system
- Booking accept/reject on schedule + color-coded cards
- Client import (CSV/contacts) + export + real earnings charts
- Booking flow: 6-step wizard with price breakdown
- Provider search: 15km radius + detail sheet
- Booking detail: status header, timeline, cancel/tip/review
- Client home: real Firestore data + upcoming bookings

+++ WoofWalk/Models/Walk.swift
@@ -6,7 +6,7 @@ struct WalkHistory: Identifiable, Codable {
     var userId: String
     @ServerTimestamp var startedAt: Timestamp?
     @ServerTimestamp var endedAt: Timestamp?
-    var distanceMeters: Int
+    var distanceMeters: Double
     var durationSec: Int
     var track: [TrackPoint]
     var polyline: String
@@ -16,7 +16,7 @@ struct WalkHistory: Identifiable, Codable {
          userId: String = "",
          startedAt: Timestamp? = nil,
          endedAt: Timestamp? = nil,
-         distanceMeters: Int = 0,
+         distanceMeters: Double = 0.0,
          durationSec: Int = 0,
          track: [TrackPoint] = [],
          polyline: String = "",
